difyを追加

+++ install.sh
@@ -12,6 +12,10 @@ find . -maxdepth 1 -type d | while read -r line; do
       ;;
     *)
       if [ -e "$name/launch.sh" ]; then
+        if [ -e "$HOME/bin/$name" ] || [ -L "$HOME/bin/$name" ]; then
+          echo "[INFO] $name はインストール済みです"
+          continue
+        fi
         echo "[INFO] $name をインストールします"
         mkdir -p "$HOME/bin"
         ln -sf "$(realpath "$line/launch.sh")" "$HOME/bin/$name"
